chore: realocate, rename, uncommit config
