+++ .gitignore
@@ -3,3 +3,4 @@ dist
 .env.local
 paybutton/dev/demo/paybutton.js
 react/coverage
+paybutton-config.json
