+++ package.json
@@ -8,6 +8,8 @@
   "license": "MIT",
   "scripts": {
     "dev": "cd paybutton && yarn dev",
+    "predev": "[ -f ./paybutton-config.json ] || cp ./example-config.json ./paybutton-config.json",
+    "prebuild": "[ -f ./paybutton-config.json ] || cp ./example-config.json ./paybutton-config.json",
     "build:dev": "yarn build:react && yarn dev",
     "test": "cd react && yarn test",
     "test:coverage": "cd react && yarn test:coverage",
