+++ react/lib/components/Widget/Widget.tsx
@@ -12,7 +12,7 @@ import copy from 'copy-to-clipboard';
 import QRCode, { BaseQRCodeProps } from 'qrcode.react';
 import io, { Socket } from 'socket.io-client';
 import PencilIcon from '../../assets/edit-pencil';
-import config from '../../config.json';
+import config from '../../../../paybutton-config.json';
 import { Theme, ThemeName, ThemeProvider, useTheme } from '../../themes';
 import { Button, animation } from '../Button/Button';
 import BarChart from '../BarChart/BarChart';
